requirements.txt
alembic==1.2.1
amqp==2.5.2
aniso8601==8.0.0
apturl==0.5.2
asn1crypto==1.0.1
Babel==2.7.0
bcrypt==3.1.7
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
cachetools==3.1.1
capstone==4.0.1
CDApplet==1.0
CDBashApplet==1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
checkbox-support==0.22
Click==7.0
command-not-found==0.3
cryptography==2.7
debtcollector==1.22.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
dnspython==1.16.0
dogpile.cache==0.8.0
eventlet==0.25.1
extras==1.0.0
fasteners==0.15
feedparser==5.1.3
fixtures==3.0.0
Flask==1.1.1
Flask-RESTful==0.3.7
futurist==1.9.0
greenlet==0.4.15
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.8
importlib-metadata==0.23
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==2.6.0
keystone-engine==0.9.1.post3
keystoneauth1==3.17.1
keystonemiddleware==7.0.1
kombu==4.6.5
language-selector==0.1
linecache2==1.0.0
louis==2.6.4
lxml==3.5.0
Mako==1.1.0
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==7.2.0
msgpack==0.6.2
netaddr==0.7.19
netifaces==0.10.9
oauthlib==3.1.0
onboard==1.2.0
os-service-types==1.7.0
oslo.cache==1.37.0
oslo.concurrency==3.30.0
oslo.config==6.11.1
oslo.context==2.23.0
oslo.db==5.0.2
oslo.i18n==3.24.0
oslo.log==3.44.1
oslo.messaging==10.2.0
oslo.middleware==3.38.1
oslo.policy==2.3.2
oslo.serialization==2.29.2
oslo.service==1.40.2
oslo.upgradecheck==0.3.2
oslo.utils==3.41.2
osprofiler==2.8.2
padme==1.1.1
passlib==1.7.1
Paste==3.2.1
PasteDeploy==2.0.1
pbr==5.4.3
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
prettytable==0.7.2
ptyprocess==0.5
pyasn1==0.1.9
pycadf==2.10.0
pycparser==2.19
pycups==1.9.73
pycurl==7.43.0
pydot==1.4.1
pygobject==3.20.0
pyinotify==0.9.6
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pysaml2==4.8.0
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.8.0
python-debian==0.1.27
python-editor==1.0.4
python-keystoneclient==3.21.0
python-mimeparse==1.6.0
python-systemd==231
pytz==2019.3
pyxdg==0.25
PyYAML==5.1.2
r2pipe==1.2.0
reportlab==3.3.0
repoze.lru==0.7
requests==2.22.0
rfc3986==1.3.2
Routes==2.4.1
scrypt==0.8.13
sessioninstaller==0.0.0
six==1.12.0
SQLAlchemy==1.3.10
sqlalchemy-migrate==0.12.0
sqlparse==0.3.0
ssh-import-id==5.5
statsd==3.3.0
stevedore==1.31.0
system-service==0.3
Tempita==0.5.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.3.0
traceback2==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unittest2==1.1.0
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.25.6
usb-creator==0.3.0
vine==1.3.0
WebOb==1.8.5
Werkzeug==0.16.0
wrapt==1.11.2
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
yappi==1.0
zipp==0.6.0
